khepri_machine: Change `get/3` to `fold/5`
The implementation of the state machine and the way we query it is
already based on walk through the tree, calling an anonymous function
when we find a matching tree node.

By exposing this to the callers of `khepri_machine`, we can provide
functions such as `khepri:fold()`, `khepri:foreach()`, `khepri:map()`
and so on. These new functions will be added in a separate commit.

`khepri_adv:get_many()` is now implemented on top of
`khepri_machine:fold/5`. It takes the responsibility of building the
node properties map.

`khepri_machine:count/3` goes away as well. `khepri:count/3` is now also
implemented on top of `khepri_machine:fold/5` and takes care of counting
tree nodes.