khepri_fun: Support the update_record/5 instruction
`update_record/5` is a new Erlang/OTP 26 instruction used to update
a field in a record with a hint about whether the data should be copied
before modification or completely re-used, along with modification
commands.

`beam_validator` must check that the register being updated is a tuple
with the given exact size (the `true` boolean in the `t_tuple` record),
so we provide an annotation for this instruction.

In order to provoke this instruction, we need a function which updates
a field in the record. The compiler must not know the full contents of
the record, though, or else it will use a `move/2` instruction instead.