Bundle keep-while into a tree record
In most cases in khepri_machine, the keep while conditions are updated
along with the root of the tree. This change bundles the root of the
tree, the keep while conditions and the keep while condition reverse
index into one record. The child commit while refactor functions that
currently take the Root and Extra to take the Tree instead and
AppliedChanges will be passed as a separate argument.