khepri_machine: Use a 'walk' record for tree-walking functions
This change refactors the `walk_down_the_tree` and related functions
to take a `walk` record which holds all of the arguments. Function
clauses in `walk_down_the_tree1` and `walk_back_up_the_tree` would
usually only use or change a subset of their parameters. The remaining
parameters would be passed along as-is, so this change cleans up and
shortens those function clauses and their callers.