Recreate ETS tables when restoring projections
When the Ra machine is restarted and enters the `recovered` state,
`restore_projection/3` re-fills the ETS table for each projection
given the current tree in the store.

This function also must recreate the ETS table for the projection,
though: if the Ra machine has snapshotted the `register_projection`
commands, they will not be replayed and the ETS tables for the
projections will not be recreated.