Store projections in a tree of patterns
The main tree structure used to store data in the Khepri store is a
tree where each edge is a path component and that tree can look up all
matching paths for a given pattern. This change introduces a new tree
structure that instead uses pattern components as edges and can be used
to look up all matching patterns for a given path.

This replaces the data structure used to hold projections in
the machine's state. The prior structure was a mapping from
projections to path patterns. Every change to the store was
checked against all patterns for registered projections using
'khepri_tree:does_path_match/3'. The existing approach is fast in
practice but it duplicates some work: pattern components shared between
multiple patterns would all be re-compiled and re-checked for each
projection and for every change to the store. The new pattern tree
eliminates the duplicate work and saves a small but noticeable amount
of time when a store uses many projections and sees a large number of
changes.