Move insert_or_update_node/5 from khepri_machine to khepri_tree
This change completes the migration started in 2ebf1cc5 which moved
functions that acted on the tree structure from 'khepri_machine' to
'khepri_tree'. With this change, the tree structure can be created and
modified without interacting with the machine state. The implementation
of 'insert_or_update_node/5' is extracted to 'khepri_tree' without
changes except that the creation of side-effects and the reorganizing
of the return value now live separately in 'khepri_machine'.

This also allows us to remove a few function exports from 'khepri_tree'
which were either being called by 'insert_or_update_node/5' or were
only used internally to 'khepri_tree'.