khepri_tree: Rename 'find_matching_nodes/5' as 'fold/5'
This change renames 'khepri_tree:find_matching_nodes/5' to the more
fitting name 'fold/5'. 'find_matching_nodes' is still a good name for
the 3-arity version of this function: it accumulates a map of the
matching nodes and their properties. However the 5-arity function is a
more generic fold function similar to 'maps:fold/3' or 'lists:foldl/3'
so the name should reflect that.