projections: Trigger changes for leaf node deletions
This change fixes a regression introduced with the refactor to the
pattern tree to hold projections for a specific case of deletions from
the tree tree.

To hit this edge-case, a path must be deleted from the tree with no
other sibling tree nodes. In other words, the deletion must not trigger
any ancestor tree nodes from being cleaned up by the default keep-while
conditions. Also, a projection must be registered to a path pattern
which would match this deleted path.

For deletion changes, we look at all of the children of the deleted
path to check if child node deletions should trigger any projections.
This block mistakenly forgot to also check if the deleted path itself
should trigger any projections.