test/cluster_SUITE: Fix 'put' node for local-reply consistency case
This part of the test case in 'cluster_SUITE' checks that
'ets:lookup/2' against a projection is immediately consistent after
causing a change with 'khepri:put/4' on the same Erlang node. The case
was mistakenly calling 'khepri:put/4' on Node1 while reading from ETS
with either Node2 or Node3 (the 'Node' binding). Both the 'put' and
'lookup' should be called from 'Node'.