Special-case trivial "copy"-style projections
A common case for projection functions is to re-use each tree node's
payload as the record for the projection ETS table. In terms of a
simple projection, this is:

    fun(_Path, Payload) -> Payload end

This change allows passing a special 'copy' atom to stand for this
case. When triggering a projection we don't execute a Horus function
so 'copy' projections have minimal overhead.