projections: Allow passing Horus options
The default Horus options used by `khepri_projection:new/3` prevent
unintentional non-determinism or interaction between processes. It could
be useful to have an escape hatch for this, though. For example you might
want to call `persistent_term:get/2` and `persistent_term:put/2` to
create some lazy value to use in the projection function.

This change allows passing in `horus:options()` when creating a
projection with `khepri_projection:new/2`. These options are merged with
the defaults and passed to `horus:to_standalone_fun/2`.