khepri_utils: Gracefully handle unavailable default applications
If the Khepri application is started and any of the default applications
(mnesia, sasl, ssl, and khepri itself) are not available,
`application:get_key/2` will return `undefined`, causing a badmatch
that will terminate the Khepri application. Instead we can gracefully
handle these applications being unavailable and not include their
modules in the list of modules to skip when extracting functions.