khepri_cluster: Make `join/3` idempotent
[Why]
If `khepri_cluster:join/3` is called twice in a row, the join succeeds
because the joining node is reset before it joins the cluster again.
This is not exactly idempotent, but the end result is the same.

However if the joining node was reset somehow (e.g. its data directory
was lost) and therefore thinks it is a standalone store, calling
`khepri_cluster:join/1` again would return `{error, already_member}`.

It is more comfortable if the function just does what it takes to
achieve the wanted result of expanding the cluster.

[How]
The `{error, already_member}` error is caught and the function returns
`ok` instead. Behind the scene, Ra did the correct thing and the joining
node is back in the cluster.