khepri_cluster: Fix how we reset a node to join a cluster
[Why]
The initial code followed these steps:
1. It removed the local Ra server from its cluster, if any. It is
   usually a no-op.
2. It force-deleted the local Ra server and deleted its data on disc.
3. It restarted it without any election.
4. It asked the remote cluster to add the local Ra server as a member.

There is one problem with them if the local Ra server was a member of
the remote cluster but it lost its data (for whatever reason). In this
case, the local Ra server didn't know about any cluster, but the remote
cluster still knew about it. When we tried to join them, depending on
the order of events, the remote cluster could send a
`append_entries_rpc` message to the local Ra server and the latter could
crash with a `leader_saw_append_entries_rpc_in_same_term` exception. The
root cause is the inconsistent view on the membership.

The remote cluster must somehow forget about the local Ra server
beforehand.

[How]
The new code does this:
1. It removes the local Ra server from the remote cluster, just in case.
   This will be a no-op in most of the cases.
2. It removes the local Ra server from its cluster, if any. This will be
   a no-op in most of the cases too.
3. It force-deletes the local Ra server and deletes its data on disc.
4. It restarts it without any election.
5. It asks the remote cluster to add the local Ra server as a member.

Basically, we remove the local Ra server from its own cluster and from
the remote cluster it wants to join. This ensures both sides are on the
same page.

While here, we do the following two changes:
* Fix a duplicate word in a log message
* Change an RPC call to a direct use of the Ra API which handles remote
  nodes just fine.