cluster_SUITE: Wait for a new leader in `can_rejoin_after_a_reset_in_a_three_node_cluster`
... after we stopped the initial leader.

[Why]
This ensures that the cluster is in a known state before we proceed with
the rest of the testcase. I got a `badmatch` once on the last line of
the following code in `ra_server.erl`:

    handle_leader({PeerId, #append_entries_reply{success = false,
                                                 next_index = NextIdx,
                                                 last_index = LastIdx,
                                                 last_term = LastTerm}},
                  State0 = #{cfg := #cfg{log_id = LogId} = Cfg,
                             cluster := Nodes, log := Log0}) ->
        ok = incr_counter(Cfg, ?C_RA_SRV_AER_REPLIES_FAILED, 1),
        #{PeerId := Peer0 = #{match_index := MI,
                              next_index := NI}} = Nodes,

We want to test that a node can be re-added to a cluster after losing
its data, but we don't want to test any corner cases like "what if the
cluster is performing an election concurrently while it still knows
about the member who lost its data".