khepri_cluster: Handle a stop of the Ra server in `do_reset()`
[Why]
When the local Ra server is to be added to a remote cluster, we perform
to member removals:
* one on the remote cluster to expand
* one on the local Ra server in case it is already in a cluster

If the local Ra server is part of the remote cluster, it will be removed
twice. That's fine. However, when we try the second removal and we have
to wait for the local Ra server to be ready for a membership change, it
may exit in case it was the leader of that cluster and is now alone.

[How]
We catch the `exit` exception when the exit reason is `normal`. In this
case, we know it was removed and exited. We don't need to try to stop it
anymore and can simply forget about the store.