khepri_machine: Run the query anonymous function from the calling process
[Why]
Before this change, the anonymous function passed to
`khepri_machine:fold/5` was executed by the Ra server process. This had
two downsides:
* it blocked the Ra server to serve writes
* the anonymous function couldn't query the Ra server itself because it
  would deadlock

[How]
The alternative implemented in this patch is to get the whole Khepri
tree from the Ra server, then run the anonymous function, this time from
the calling process.

The downside is that the whole Khepri tree is copied to that calling
process.