khepri_machine: Use cached leader for async commands
Async commands will fail to be applied unless they're sent to the
cluster's current leader. With this change we always try to use the
cached leader from synchronous commands if available and use the local
member if the leader isn't cached. The caller can then update the cached
leader value when the async response says that the command was sent
against a non-leader member, like so:

    receive
        {ra_event,
         FromId,
         {rejected, {not_leader, MaybeLeader, CorrelationId}}} ->
            ok = khepri_cluster:cache_leader_if_changed(
                   StoreId, FromId, MaybeLeader),
            %% ... retry the command for `CorrelationId' ...
            ok
    end