Add khepri:handle_async_ret/2 for handling async Ra events
This is similar to `wait_for_async_ret/1,2` except that the caller is
expected to `receive` the Ra event. This is necessary because the Ra
event sent when a batch of commands is applied includes a list of
correlation and result pairs, making it impossible to use a selective
receive to look for a single correlation ID.

We could alternatively map the `ra_event` message into some other value
like `{ok, Correlations} | {error, CorrelationId}` but this isn't much
easier to work with than the `ra_event`s themselves and would be brittle
for changes to `ra_event`. Instead we leave it to the caller to receive
and handle the events. This function is then responsible only for
calling internal functions to update the cached leader for the given
store ID.