Remove khepri:wait_for_async_ret/1,2 in favor of handle_async_ret/2
See the parent commit: `wait_for_async_ret/1,2` used a selective receive
to find whether a given correlation ID was applied. The `ra_event`
emitted when a batch is applied gives all correlation and result pairs
applied in the batch though, making it impossible to reliably find a
single correlation ID. We should instead have the caller perform the
`receive` to find the applied/rejected event and handle all of the
correlation IDs found within.

So this change removes `wait_for_async_ret/1,2` in favor of the
`handle_async_ret/2` function added in the parent commit.