include/khepri.hrl: Expose `?IS_KHEPRI_STORE_ID/1` macro
[Why]
This allows to abstract the type of the store ID.

[How]
The macro already existed for a long time, but was in a private header.
This patch moves it to the public header.

In the process, it is renamed from `?IS_STORE_ID/1` to
`?IS_KHEPRI_STORE_ID/1` to reduce the namespace pollution.