export_erlang: Use cwd for fail_to_open_file_test_
Using root for the filename fails for me locally on a Mac laptop. I
guess it must have something to do with the read-only root feature -
there is some sort of misdirection between `/` and `/private`. This case
fails on my machine but with `{error,eexist}` rather than the expected
`{error,eisdir}`. Using the current working directory fails as expected.