khepri: Always pass exported functions to Ra queries
[Why]
Using `fun()` is fragile if it has to be executed on a remote node.
Indeed, the remote node may have another version of the module hosting
the function and the function reference may be invalid.

[How]
Now, we always use an exported function.

Callers of any folding APIs should do the same.

Fixes #238.