khepri: Ensure queries are passed as MFA if they can run remotely
[Why]
Using `fun()` is fragile if it has to be executed on a remote node.
Indeed, the remote node may have another version of the module hosting
the function and the function reference may be invalid.

[How]
In the previous commit, Khepri stopped using fun references internally
where we know they can be executed on a remote node.

In this commit, we ensure that we apply the same constraint to external
functions passed to fold/foreach/map/filter queries and read-only
transactions.

If the caller passes such a function as a function reference instead of
the MFA tuple, Khepri will throw an exception to indicate this is not
permitted.

Here is an example for those queries and R/O transactions:

    khepri:fold(
      StoreId,
      PathPattern
      {my_mod, my_fold_func, [SomeState]},
      InitialAcc).

Where `my_mod:my_fold_func()` looks like:

    my_fold_func(SomeState, Path, NodeProps, Acc) ->
        ...
	NewAcc.