khepri_cluster: Functions returning members now wrap them in `{ok, _}`
[Why]
Always returning a list, hiding any error behind an empty list, proved
to be a bad design, as it made it really difficult to detect errors and
act upon them.

This kind of return value was also inconsistent with the rest of the
API.

[How]
The `members/{0,1,2}`, `locally_known_members/{0,1,2}`, `nodes/{0,1,2}`
and `locally_known_nodes/{0,1,2}` functions now return `{ok, List}` or
`{error, Reason}`, like other APIs.

The returned list is never empty.

While working on this, the variants taking no arguments were added to
work with the default store.