khepri_machine: Use monotonic time, not system time
[Why]
System time is not time warp safe, thus it could go backward.

[How]
We replace the uses of `erlang:system_time()` by
`erlang:monotonic_time()`.

While here, we get and store the monotonic time in native units, and
only convert the time diff to seconds. This reduces the number of
conversions from 2 to 1 and it improves the precision a bit.