khepri_machine: Do not enforce `favor` options in machine state queries
[Why]
We want to leave the caller decide the consistency it wants.

Also, there is a problem with leader and consistent queries in Ra
because the function is executed on a remote node and the function
reference or the MFA may not be valid, leading to an exception.

[How]
We just stop setting the `favor` option.