Fix unregistering non-existent projection when pattern tree is non-empty
Attempting to unregister a projection that doesn't exist failed when
the `khepri_pattern_tree` of projections was non-empty (i.e. there was
some other projection registered) because the call to
`khepri_pattern_tree:filtermap/2` returned a value that breaks the
`filtermap/2` spec. Fixing the return value fixes the behavior of
unregistering the non-existent projection.