cluster_SUITE: Add test to ensure commands are not duplicated
[Why]
When `ra:process_command()` returns an error like `{error, shutdown}`,
it doesn’t mean the command was nod appended to the Ra log. It’s to be
expected with networks, but something I forgot in Khepri.

The retry mechanism we have will call `ra:process_command()` again after
specific errors. Therefore, a command may be appended and thus applied
twice. We observed exactly this in a testcase of the rabbitmq-shovel
plugin.

To help us work on a solution, we need a smaller testcase.

[How]
This new testcase reproduces the problem by spamming a Khepri store with
transactions that bump a counter, while the cluster is changed to
trigger an election.

At the end, the counter should be equal to the number of times the
transaction was executed.