khepri_machine: Add command deduplication mechanism
[Why]
The "client" side of `khepri_machine` implemented in
`process_sync_command/3` have a retry mechanism if
`ra:process_command/3` returns an error such as `noproc`, `nodedown` or
`shutdown`.

However, this retry mechanism can't tell if the state machine already
received the command and just couldn't reply, for instance because there
is a node stopping or a change of leadership.

Therefore, it's possible that the same command is submitted twice and
thus processed twice.

That's ok for idempotent commands, but it may not be alright for all
transactions for example. That's why we need a deduplication mechanism
that ensures the same command is not applied multiple times.

[How]
Two new commands are introduced to implement the deduplication system:
* #dedup{} which is used to wrap the command to protect and assign a
  unique reference to it
* #dedup_ack{} which is used at the end of the retry loop to let the
  state machine know that the "client" side received the reply

When the state machine receives a command wrapped into a #dedup{}
command, it will remember the reply for the initial processing of that
command. For any subsequent copies of the same #dedup{} (based on the
unique reference), the state machine will not apply the wrapped command
and will simply returned the reply it remembered from the first
application.

Later when the state machine receives a #dedup_ack{}, it will drop the
cached reply for that reference.

Just in case the client never sends a #dedup_ack{}, the state machine
will drop any expired cached entries. The expiration time is based on
the command timeout. If it's infinity, it defaults to 15 minutes.

This whole deduplication mechanism can be enabled or disabled through
the new `protect_against_dups` command option which takes a boolean.
This option is off by default, except for R/W transactions.

Thus if the caller knows the transation is idempotent, it can decide to
turn the dedup mechanism off.

Because the state machine's state grows with a new field and handles two
new commandes, we bump the machine version from 0 to 1.

V2: We now use the `effective_machine_version` counter provided by
    `ra_counters:counters/2` if it is available as it is faster than
    querying the Ra server. If the counter is unavailable, we fall back
    to the query. The new counter is added by rabbitmq/ra#426 and will
    be used once a Ra release contains this change.