khepri_machine: Handle unknown state machine commands
[Why]
In the future, we will introduce new commands that older versions of the
state machine won't know about.

Before this commit, the state machine would crash with a function clause
exception. This could be lead to a Ra cluster having a hard time to
recover if all Ra servers crash.

[How]
If the machine is asked to apply an unknown command, it returns an
error. This error is converted to an exception in
`khepri_machine:process_command()`.

While here:
* Add `machine_version` and a valid `system_time` value to the `?META`
  macro used by tests.
* Fix an invalid log message where arguments to `?LOG_DEBUG()` were
  incorrect.

V2: Log an error using a `mod_call` side effect, not directly from the
    state machine process using `?LOG_ERROR`. The former solution would
    log on all Ra cluster members and even during recovery. The
    `mod_call` side effect is only processed by the leader.