Make #khepri_machine{} private
[Why]
With the upcoming need to add a field to implement command
deduplication, we will have two versions of the record for the machine
state.

If we expose the definition, it will be more complicate to track which
one we are using.

[How]
The record definition is moved to `src/khepri_machine.erl` and made an
opaque type.

The module exports functions to access the fields of the record without
ever exposing the record itself. Most of these functions were already
exported if `TEST` was defined at compile-time. This is no longer
needed.