khepri_machine: Use functions to manipulate the internal state
... instead of using the record directly.

[Why]
This is a preparation step for a future update of the record. At that
point, there will be two versions of the state record to support and
it's easier to isolate that handling in a few functions.

[How]
All reads and writes to the state go through a set of `get_*/1` and
`set_*/2` functions.