Add khepri:has_projection/1,2,3
This function checks if a projection is currently registered. A caller
might want to use it instead of sending `#register_projection{}`
commands and relying on idempotency because `#register_projection{}`
commands contain a horus standalone function, and the horus function's
beam code can consume a non-trivial amount of data in the Ra log.

We will use the `khepri_machine:has_projection/2` helper function in
the child commit to check whether a projection exists within the
`apply/3` clause that handles `#register_projection{}` commands so that
we can check for existing projections without the side effect of
creating an ETS table.