khepri_machine: Use an effect to initialize new projections
Previously the machine would initialize projections in-place within
`apply/3`. That would cause projections to be restored ahead of time
when the machine was recovering if it applied a `#register_projection{}`
command.

We can use an aux effect instead for the creation and initial filling
of the projection's ETS table. Aux effects are skipped during recovery.
Once the machine is fully recovered and the server enters the
`recovered` state, we emit a `restore_projections` aux effect that
properly (re)creates and fills the projections. (This is already done by
`khepri_machine`.)