khepri_pattern_tree: Fix conditions that were matching subtrees
[Why]
A pattern tree with `/parent/*` is supposed to match
`/parent/child` but not `/parent/child/grandchild`.

Before this patch, this was the case because we didn't check if we
reached the of the path before calling the folding function in
`khepri_pattern_tree:fold1/6`.

[How]
Now, we check the context and only call the fold function if we reached
the end of the path.

Otherwise, we recurse without calling that fold function, checking a few
more things depending on whether the condition could match grand
children or not.

The testsuites are expanded with more testcases to cover these
situations.