khepri_machine: Queries default to local ones
[Why]
Before, Khepri tried something smart by mixing consistent and leader
queries by default. It exposed the `favor` option to let the caller
influence the type of query.

Leader queries are a good compromise in terms of latency and freshness
of the queried data. However, they come with a big caveat: the query
function is executed on the leader, not on the local node (if the local
node is a follower).

There are various potential issues with the remote execution of query
functions:
* The function reference might be invalid on the remote node because of
  a different in the version of Erlang/OTP.
* The function reference might be invalid on the remote node because the
  module hosting the function is not the same.
* The function reference might be invalid on the remote node because the
  module hosting the function is missing.
* Using an exported MFA instead of a function reference doesn't solve
  the problem because the remote copy of that MFA may still be missing
  or simply be incompatible (different code/behavior). This might be
  even more difficult to debug.

[How]
The only way to be sure that the query function behaves as the caller
expects is to the local copy of that function. Therefore, we must use a
local query.

A local query has its own limitations. In particular it could work with
out-of-date data. That is why a previous commit changed the default
behavior of synchronous updates so that the call returns only when it
was applied locally.

Therefore, with the new default options in this commit and the previous
ones, we ensure that a query will work with the local query function and
that it will be executed against the local up-to-date state.

There is one more corner case: the use of asynchronous updates (that use
Ra pipelined commands). See the next commit for that.