khepri_machine: Introduce a fence mechanism
[Why]
With the previous commits, we ensured that when a synchronous update
returns, the local machine state is up-to-date. The chance that it is
the case for async commands without a correlation ID has also increased.

In a follow-up commit, Khepri will default to local queries, to
eliminate the risks linked to remote executions and possible
incompatible code.

Therefore, a query is local to make sure its execution is local and that
it works on an up-to-date state.

There is still the following scenario where this isn't enough:
1. The caller makes several asynchronous updates without a correlation
   ID, or overrides the default of `reply_from => {member,
   LocalRaServer}`.
2. The caller then performs a query.

In this case, the query is unlikely to be executed after the
asynchronous commands were applied. This is not a bug, the caller
explicitly asked for asynchronous updates.

The caller could use correlation IDs and wait for the replies. But
without correlation IDs, it's not possible.

[How]
To help the caller in this case, this patch introduces
`khepri:fence/{0,1,2}`. It is a blocking call that queries the Ra leader
to learn its last index (the number of the last command it received),
then performs an arbitrary local query, passing that index so that the
query execution waits for that index to be committed locally.

This way, by putting a call to `khepri:fence()` between asynchronous
updates and a query, the caller ensures that the query will see the
result of those asynchronous updates.

To make this mechanism effective, we also send all synchronous commands
and asynchronous commands without a correlation ID to the local Ra
server, instead of the leader even if we know who it is. This way, all
messages are "serialized".