khepri_machine: Commands default to `reply_from => local`
[Why]
The default behavior of Khepri is that a reply to a synchronous command
is sent to the caller once a quorum number of Ra members committed the
command and the leader applied it.

This list of Ra members might not include the Ra server local to the
caller. If the caller then depends on the local state, for local queries
or projections, it might not see the update that was successfully
applied yet. This behavior might be surprising.

The default is now to send the reply to a synchronous command only once
a quorum number of Ra members committed the command and the local Ra
server applied it, regardless it is the loader or a follower. This
ensures the local state is in sync with whatever the caller might expect
after a successful update.

[How]
The default value of the `reply_from` option is simply set to `local`.

The caller might still decide to change it for a smaller latency.

This change is also required by the following changes to queries.