khepri_machine: Asynchronous commands default to the `normal` priority
[Why]
The default priority is `low` in Ra. This means that the leader may
batch all low priority commands and only commit them after some internal
conditions.

For Khepri, we change that default to `normal` to have a behavior closer
to synchronous commands.

This brings a more logicial behavior to queries, or more exactly a less
surprising one. Indeed, we want to push for local queries by default and
add a "fence" mechanism in subsequent commits for various reasons (see
the commit messages to learn more). Having asynchronous commands that
behave more closely to synchronous ones makes it easier to reason about
the caveats around queries and the local Ra server state.