khepri_cluster: Don't cache the `not_known` "leader"
[Why]
Local or leader queries may return `not_known` instead of the leader ID
if this one isn't known as part of the query.

So far, we cached that value and this value was used as the leader ID in
follow-up calls to Khepri... This is incorrect as this is clearly not a
valid Ra server ID.

[How]
If the passed "leader" ID is `not_known`, we ignore it and don't cache
anything.

We don't clear the cached leader either because we lack context to take
that decision.