khepri: Add leader ID to the `not_leader` error tuple
... in `handle_async_ret/2`

[Why]
If the async command with a correlation ID is sent to a follower, that
follower will reply with the `not_leader` error. The caller is
responsible for resending the command to the leader that is specified in
the `not_leader` error.

[How]
This piece of information was not passed to the caller by
`handle_async_ret/2` so the caller couldn't do it. The API is fixed by
this commit.