khepri_machine: Send async commands to the local Ra server
... if possible.

[Why]
This will be useful for the upcoming fence mechanism (see a later
commit). This will ensure that commands, even async ones, and queries
are managed in order.

[How]
For async commands without a correlation ID, we can send them to the
local Ra server. It will handle redirection to the leader if the local
server is a follower.

For those with a correlation ID, the local Ra server won't handle
redirections. Therefore, we must query the leader and send the command
to it. If we don't know the leader, it is sent to the local Ra server
and if it is a follower, it will reply with an error. Async commands
with a correlation ID won't be covered by the fence mechanism.