khepri_machine: Don't set `reply_from` when (un)registering projections
[Why]
`reply_from` is ineffective in the first place in this case because
`process_command/3` could have sent the command to the leader directly.

[How]
Leave `process_command/3` set `reply_from`. By default, it will use the
local member ID. This is the same intention than `local` but it will
work regardless of the node that the command is sent to.