cluster_SUITE: Fix `handle_leader_down_on_three_node_cluster_command/1` testcase
[Why]
The testcase was issueing a put on node 1, then a get on all nodes. From
time to time, one of the gets would fail because the queried node didn't
see the put yet.

[How]
Now, we use the `#{favor => consistency}` option for gets on nodes that
are neither the leader, nor the follower that issued the put.

Indeed, to increase the testing of consistency, we also perform the put
from a follower instead of node 1.

This is the same fix as committed in commit a076a83 (#269) for another
testcase.