khepri_cluster: Clean running stores list up in `is_store_running/1
[Why]
If a Ra server has stopped behind the back of Khepri, we want to delete
it from the list of running stores.

[How]
Instead of having an assertion to ensure that the presence or absence of
a store in the list of running stores matches the result of
`is_store_running/1`, we clean up the list if needed as part of that
function.