khepri_export_erlang: Accept atoms and binaries as filenames
[Why]
The `file:name_all()` type accepts atoms and binaries in its definition.

Also, Elixir strings are binaries really, so using `"filename.erl"` as
the filename would break because the module would consider that binary
filename as an already opened file descriptor.

[How]
The guard expression is expanded to accept atoms and binaries.

Fixes #278.